+++ docs/Changelog.md
@@ -32,7 +32,8 @@
       AFL type forkserver. Useful for symcc/symqemu/nautilus/etc. with
       AFL_LLVM_INSTRUMENT=CLASSIC
   - code formatting updated to llvm 18
-  - improve /custom_mutators/aflpp/standalone/aflpp-standalone
+  - improved custom_mutators/aflpp/standalone/aflpp-standalone
+  - added custom_mutators/autotokens/standalone/autotokens-standalone
 
 
 ### Version ++4.21c (release)